Model Modification
Added "comment dependent destroy" to Article so that when an article
is deleted, the comments are deleted as well.

Added "comment dependent destroy" and "article dependent destroy" to User
so that if a user's account is deleted, their articles/comments are
deleted as well

For both the Article and Comment Serializer, date_created was switched
to updated_at, so that the timestamps can show the most previous activity.